fix(tool/browser): address review feedback for CDP browser tool
Critical fixes:
- Fix screenshot file lifecycle: don't delete tmpFile after MediaStore.Store
  (MediaStore only records a mapping, removing the file breaks the media ref)
- Add stateMu for concurrent-safe access to history/mediaStore/cdp state
- Fix mutex ordering race: executeClose now acquires cdpMu before stateMu
- Harden SSRF: block numeric hostnames (0x7f000001, 127.1, 2130706433),
  CGNAT (100.64.0.0/10), benchmark (198.18.0.0/15) ranges
- Document DNS rebinding limitation inherent to CDP architecture

Other fixes:
- Fix scrollIntoView behavior: 'instant' -> 'auto' (Chrome 125+ TypeError)
- Handle json.Unmarshal errors in executeType/executeFill (was silently ignored)
- Thread context through CDP SendCtx for proper timeout/cancellation
- Clear stale data-pcw-idx attributes in executeState to prevent stale clicks
- Fix env var naming: PICOCLAW_TOOLS_BROWSER_TIMEOUT -> _TIMEOUT_SECONDS
- Fix stub Parameters() schema: add "required" field for consistency
- Fix NewBrowserTool doc comment: connection is lazy, not eager
- Fix screenshot integration test assertion to match no-MediaStore path
- Split browser.go (990 lines) into browser.go + browser_actions.go (<600 each)
- Add 8 new SSRF bypass test cases
